// src/store/Country/country.action.type.ts
export enum Country {
  SET_LOADING = "SET_LOADING",
  SET_ERROR = "SET_ERROR",
  CLEAR_ERROR = "CLEAR_ERROR",
  GET_COUNTRY_DATA = "GET_COUNTRY_DATA",
  GET_SINGLE_COUNTRY_DATA = "GET_SINGLE_COUNTRY_DATA",
  GET_BORDER_COUNTRY_NAMES = "GET_BORDER_COUNTRY_NAMES",
  CLEAR_SINGLE_COUNTRY_STATE = "CLEAR_SINGLE_COUNTRY_STATE",
  CLEAR_COUNTRIES_STATE = "CLEAR_COUNTRIES_STATE",
  CLEAR_BORDER_COUNTRIES_STATE = "CLEAR_BORDER_COUNTRIES_STATE",
  GET_COUNTRY_DATA_BY_SEARCH = "GET_COUNTRY_DATA_BY_SEARCH",
  GET_COUNTRY_DATA__BY_CONTINENT = "GET_COUNTRY_DATA__BY_CONTINENT",
}